pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
# Project Metadata
# ================
name = "chem_mat_database"
version = "1.1.2"
description = "Command Line Interface for projects"
readme = "README.rst"
authors = [
    {name = "Jonas Teufel", email = "jonseb1998@gmail.com"},
    {name = "Mohit Singh", email = "mohit.singh@student.kit.edu"}
]
maintainers = [
    {name = "Jonas Teufel", email = "jonseb1998@gmail.com"},
]
license = { text = "MIT License" }
keywords = [
    "chemistry",
    "data",
    "dataset",
    "machine learning",
    "data management",
    "graph neural network",
    "graph representation",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

# Project Dependencies
# ====================
requires-python = ">=3.8,<=3.12"
dependencies = [
    "numpy>=1.21.0,<2.0",
    "matplotlib>=3.4.0",
    "rdkit>=2023.0.0",
    "ase>=3.23.0,<4.0",
    "requests>=2.30.0",
    "cairosvg>=2.7.1",
    "imageio>=2.34.0",
    "rich_click>=1.8.0",
    "jinja2>=3.0.0",
    "lorem==0.1.1",
    "networkx<=3.3",
    "python-dotenv==1.0.1",
    "pyyaml>=6.0.0",
    "msgpack>=1.0.2",
    "pandas<2.2.2",
    "pycomex<=0.17.0",
    "msgpack_numpy>=0.4.8",
    "appdirs>=1.4.4,<2.0",
    "tomlkit>=0.13.2,<1.0",
    "openpyxl>=3.0.10,<4.0",
    "prettytable>=3.6.0,<4.0",
    "pydantic>=1.10.0,<2.0",
    "parsimonious>=0.10.0,<1.0",
]

# Optional Dependencies
# =====================

[project.optional-dependencies]
dev = [
    # Testing dependencies
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",

    # Development tools
    "nox[uv]>=2023.4.22",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-requests>=2.31.0",

    # ML/Graph dependencies
    "torch>=1.13.0",
    "torch-geometric>=2.3.0",
    "jax>=0.4.0",
    "jaxlib>=0.4.0",
    "jraph>=0.0.6.dev0",

    # Documentation
    "sphinx>=6.0.0",
    "sphinx-rtd-theme>=1.3.0",

    # Build tools
    "build>=0.10.0",
]

# Executable Scripts
# ==================

[project.scripts]
cmdata = "chem_mat_data.cli:cli"
cmmanage = "chem_mat_data.manage:cli"

# Configuration of Build System (Hatchling)
# =========================================

[tool.hatch.build]
skip-excluded-dirs=true

[tool.hatch.build.targets.wheel]
packages = ["chem_mat_data"]

# Version Bumping
# ===============

[tool.bumpversion]
current_version = "1.1.2"
parse = "(?<![=<>])(?P<major>\\d+)\\.(?P<minor>\\d+)\\.(?P<patch>\\d+)"
serialize = ["{major}.{minor}.{patch}"]
search = "{current_version}"
replace = "{new_version}"
regex = true
ignore_missing_version = true
tag = false
sign_tags = false
tag_name = "v{new_version}"
tag_message = "Bump version: {current_version} → {new_version}"
allow_dirty = true
commit = false
message = "Bump version: {current_version} → {new_version}"
commit_args = ""
files = [
    {filename = "pyproject.toml", search="version = \"{current_version}\"", replace="version = \"{new_version}\""},
    {filename = "README.rst"},
    {filename = "chem_mat_data/VERSION"},
]